feature/Conversao

+++ src/main/java/com/fabriciosanches/fichatecnica/dtos/ConversaoValoresDTO.java
@@ -2,6 +2,6 @@ package com.fabriciosanches.fichatecnica.dtos;
 
 import java.math.BigDecimal;
 
-public record ConversaoValoresDTO(Integer quantidade, Long unidadeMedidaPara,
+public record ConversaoValoresDTO(Double quantidade, Long unidadeMedidaPara,
                                   BigDecimal valor) {
 }
